Replace ragnarok online assets
Changed the assets for mage and knight from ragnarok online assets to ones provided free of charge off itch.io.
Changes how the sprites for the knight and mage classes are loaded, using an ImageGrid, which is like a sprite sheet, over individual files.
Certainly saves me a bunch of time slicing the images up myself.