Tweak absolutely positioned items and camera
I was playing with the camera's origin and the window size, and realized most "fixed" sprites, such as menu items, depend greatly on where the camera's origin was placed.
The `offset_x` and `offset_y` properties of the camera returned values that, when used with the current magic constants for menu items, looked great.
Once I shifted the starting camera origin up 100px though, everything was unusable.

Moves the logic of what "absolute" pixels are to the camera class.
It exposes methods for taking an X and/or Y that we want from the bottom left of the window and converting that to whatever makes sense in absolute X and Y based on where the camera is.

Granted, all of this will only end up working when the camera does not move.
All the sprite objects are fixed in space, and if the camera moves, say with the current player's turn, you have to move the objects to follow.
Ok for now, probably a much better place absolute positioned UI should end up living.

I think some of this work helps move us towards a bigger window resolution.
There are a bunch of magic positions baked into the code though, including what I am now committing for the camera conversion methods.