Make tab key switch between characters in game
Adds a function which moves the highlight to the next available character on the team of the current turn.
Calls this function on the scene's first render, moving the cursor from 0,0 to the first character.
Also adds a binding for the tab key during select mode to help cycle between characters.