Rename characters & give map control of starting positions
Renames the characters to something more akin to their stats and play style than their previous graphics.

Adds a utility function to the map class to generate 4 lists of size `team_size`
These lists will contain a row, a column and a direction tuple.
The data represents starting location information for up to 4 teams on the map.
The direction tells us what way the character should face to look at the center of the map, which therefore has all the teams looking at their direct opponent.