Display health on characters
Moves the health tracking from characters to a Health class
This class is drawable, and displays as `X/Y`, and appears above the characters.
To keep this from being overlapped by other characters, we draw the characters first and then the UIs after.

Also moves a touch of the combat logic into the character / health class.
Much of what's going on in the GameScene should be delegated.