Move map responsibilities to Map class
This started as trying to replace the isotiles/grass images with a different tile.
When I swapped in the image from block, and the grid just didn't display properly at all, I realized I needed to modify the map generation in the GameScene.
I already knew the GameScene had too much responsibility, and trying to change the code for map generation sealed the deal.

- Removes the isotiles.gif and grass.png files, and the Grass character
- Adds blocks.png, isometric 64x64 tiles which will help add more detail to the map in the future
- Renames Map to RectangularMap.
  Gives the class more responsibility around rendering itself, drawing the cursor and highlighting movement/attack radii
- Changed references to height with depth when referring to the map, as actual z-height on a map may be coming soon

The math around the sprite size and how sprites are layed out properly to overlap and built up the map is documented in comment in the RectangularMap class.
Right now, the 64x64 tile has it's center point at 32px across and 16px down.
This is hard coded into the map class.
I believe this can be pulled out one more layer into a data tuple of sprite and offsets, which would help make the map class tile agnostic.

The characters are a touch too large for these tiles.
I'll address that in another fix.