Fix map sprite "outlining"
When the camera first moves in the game scene, all the map tiles get a "grey" outline to them.
After googling, I believe this is anti-aliasing kicking in on the textures with alpha channels.
The textures are being placed at sub-pixel locations as the camera re-focuses.

Using glTexParameteri as suggested by a few stack overflows.
This is definitely semi-magic to me at the moment.
I believe the correct fix is to understand the openGL command that are setting up perspective and moving the modelview matrix a little more.
Making the matrix transformations in whole pixels may be the correct fix.