Force camera to move in whole numbers
Changed the targeting and ticking logic to have the camera always pointing at a whole number X and Y.
This will make it so that the sprites are never lined up on sub-pixels, and therefore, never have to anti-alias.
With this change, we no longer need the glTexParameteri calls in the map draw method.